ISSUE-1: - Renamed `LoggerHandler` into `LoggerCollection`.
