+++ README.md
@@ -13,13 +13,13 @@ Bundle provides the ability to connect loggers registered in `MonologBundle` thr
 The goal is achieved thanks to auto-generated classes of loggers. Each class decorates one object of one
 of the registered `monolog` channel.
 
-Also available is the second way to achieve the goal - using the control class `LoggerHandler`.
+Also available is the second way to achieve the goal - using the `LoggerCollection`.
 If the requested channel does not exist - will be selected fallback `logger`.
 As fallback `logger`  will be used service, which referenced by `@Psr\Log\LoggerInterface`.
 In that case where `logger` was not registered in service container will be returned instance of `Psr\Log\NullLogger`.
 
 **Important:** Bundle will work properly in the absence of `MonologBundle` in the project.
-In that case `LoggerHandler` will always return a fallback value.
+In that case `LoggerCollection` will always return a fallback value.
 
 Installation
 -----------
@@ -141,33 +141,46 @@ class AcmeLoggerAware
 }
 ```
 
-### Use through LoggerHandler
+### Use through collection of loggers
 
 ```php
 <?php declare(strict_types=1);
 
-use Linkin\Bundle\MonologAutowireBundle\Handler\LoggerHandler;
+use Linkin\Bundle\MonologAutowireBundle\Collection\LoggerCollection;
+use Psr\Log\LoggerInterface;
 
 class AcmeLoggerAware
 {
     /**
-     * @var LoggerHandler
+     * @var LoggerInterface
      */
-    private $loggerHandler;
+    private $acmeLogLogger;
 
     /**
-     * @param LoggerHandler $loggerHandler
+     * @var LoggerInterface
      */
-    public function __construct(LoggerHandler $loggerHandler) 
+    private $doctrineLogger;
+
+    /**
+     * @var LoggerInterface
+     */
+    private $logger;
+
+    /**
+     * @param LoggerCollection $loggerCollection
+     */
+    public function __construct(LoggerCollection $loggerCollection) 
     {
-        $this->loggerHandler = $loggerHandler;
+        $this->acmeLogLogger = $loggerCollection->getLogger('acme_log');
+        $this->doctrineLogger = $loggerCollection->getLogger('doctrine');
+        $this->logger = $loggerCollection->getLogger();
     }
     
     public function doSome(): void
     {
-        $this->loggerHandler->getLogger('acme_log')->info('INFO into "acme_log" channel');
-        $this->loggerHandler->getLogger('doctrine')->info('INFO into "doctrine" channel');
-        $this->loggerHandler->getLogger()->info('INFO into Fallback or into NullLogger');
+        $this->acmeLogLogger->info('INFO into "acme_log" channel');
+        $this->doctrineLogger->info('INFO into "doctrine" channel');
+        $this->logger->info('INFO into Fallback or into NullLogger');
     }
 }
 ```
